Remove Router & Profile
Although a router would probably make this a better app, my goal is
to keep this app lean and simple and easy to maintain. So this PR
removes the router entirely, as well as all Profile-related items.

It also makes a few CSS improvements to the header & footer.

Fixes #48
