+++ src/components/BookList/BookList.scss
@@ -7,10 +7,6 @@
   padding: 0;
 
   @media screen and (min-width: 640px) {
-    grid-template-columns: 1fr 1fr;
-  }
-
-  @media screen and (min-width: 1024px) {
     grid-template-columns: repeat(auto-fill, minmax(375px, 1fr));
   }
 }
