+++ src/components/TheFooter/TheFooter.scss
@@ -1,0 +1,3 @@
+.footer {
+  margin-top: 1.5rem;
+}
