+++ src/components/TheHeader/TheHeader.scss
@@ -3,6 +3,7 @@
   display: flex;
   flex-direction: column;
   justify-content: space-between;
+  margin-bottom: 1rem;
 
   > * + * {
     margin-top: 0.5em;
@@ -19,6 +20,7 @@
 
 .header__logo {
   font-weight: 400;
+  line-height: 1;
   margin: 0;
 
   @media screen and (min-width: 768px) {
