--- src/pages/Home.vue
@@ -1,10 +1,0 @@
-<template>
-  <MainLayout>
-    <BookList />
-  </MainLayout>
-</template>
-
-<script setup>
-import MainLayout from '../layouts/Main.vue';
-import BookList from '../components/BookList/BookList.vue';
-</script>
