--- src/pages/NotFound.vue
@@ -1,9 +1,0 @@
-<template>
-  <MainLayout>
-    <p>Page not found.</p>
-  </MainLayout>
-</template>
-
-<script setup>
-import MainLayout from '../layouts/Main.vue';
-</script>
