--- src/pages/Profile.vue
@@ -1,10 +1,0 @@
-<template>
-  <MainLayout>
-    <ProfileEdit />
-  </MainLayout>
-</template>
-
-<script setup>
-import MainLayout from '../layouts/Main.vue';
-import ProfileEdit from '../components/ProfileEdit/ProfileEdit.vue';
-</script>
